eic_case_studies/requirements.txt
numpy
pandas
boto3
requests
pyyaml